create/modify with or without content, subject or image

+++ backend/controllers/posts.js
@@ -10,7 +10,11 @@ const con = mysql.createConnection({
 
 exports.createPost = (req, res, next) => {
   const postObject = req.body;
-  postObject["media"] = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  if (req.file) {
+    postObject["media"] = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } else {
+    delete postObject.image
+  }
   console.log(postObject)
   con.query('INSERT INTO post SET ?', postObject, (err, resp) => {
     // console.log(resp)
@@ -70,11 +74,55 @@ exports.modifyPost = (req, res, next) => {
     if (req.params.userId == resp[0].creatorId || req.params.role == "modo") {
       const postObject = req.body 
       console.log(req.file) 
-      const filename = resp[0].media.split('/images/')[1];
       if (req.file) {
-        const filename = resp[0].media.split('/images/')[1];
         postObject["media"] = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-        fs.unlink(`images/${filename}`, () =>{
+        if (resp[0].media) {
+          const filename = resp[0].media.split('/images/')[1];
+          fs.unlink(`images/${filename}`, () =>{
+            let quer = ""
+            let inc = 0
+            let final_range = Object.keys(postObject).length
+            for (field of Object.keys(postObject)) {
+              inc += 1
+              if (field!="id"){
+                if (inc == final_range) {
+                  quer = quer.concat(field,`=${mysql.escape(postObject[field])} `)
+                } else {
+                  quer = quer.concat(field,`=${mysql.escape(postObject[field])}, `)
+                }
+              }
+            }
+            console.log(quer, resp[0])
+            con.query(
+              `UPDATE post SET ${quer} Where id = ?`, req.params.postId, (err, resp) => {
+                if (err){res.status(400).json({ err })}
+                else {res.status(201).json({ message: 'Objet modifié !'})}
+              })
+          })
+        } else {
+          let quer = ""
+            let inc = 0
+            let final_range = Object.keys(postObject).length
+            for (field of Object.keys(postObject)) {
+              inc += 1
+              if (field!="id"){
+                if (inc == final_range) {
+                  quer = quer.concat(field,`=${mysql.escape(postObject[field])} `)
+                } else {
+                  quer = quer.concat(field,`=${mysql.escape(postObject[field])}, `)
+                }
+              }
+            }
+            console.log(quer, resp[0])
+            con.query(
+              `UPDATE post SET ${quer} Where id = ?`, req.params.postId, (err, resp) => {
+                if (err){res.status(400).json({ err })}
+                else {res.status(201).json({ message: 'Objet modifié !'})}
+              })
+        }
+      } else {
+          delete postObject.image
+          delete postObject.id
           let quer = ""
           let inc = 0
           let final_range = Object.keys(postObject).length
@@ -90,12 +138,10 @@ exports.modifyPost = (req, res, next) => {
           }
           console.log(quer, resp[0])
           con.query(
-            `UPDATE post SET ${quer} Where id = ?`,
-            req.params.postId, (err, resp) => {
+            `UPDATE post SET ${quer} Where id = ?`, parseInt(req.params.postId), (err, resp) => {
               if (err){res.status(400).json({ err })}
               else {res.status(201).json({ message: 'Objet modifié !'})}
             })
-        })
       }
     } else {
         res.status(400).json({message: "You are neither the creator or the moderator of this post."})
